Refresh sources mtime after using them
This is due to a particularity of UFSCar's cluster, so we only do
this in the Singularity backend. However, it might not be required
in other HPC clusters. UFSCar's cluster deletes old files from the
scratch storage automatically. Problem is bare VCS repositories
have different mtime for different files, thus some files might
get deleted whereas other files of the same directory remain.
This effectively corrupts the VCS repository and causes makepkg
to fail. By refreshing mtime, we notify systemd-tmpfiles the
files were recently used and, moreover, cause all files to be
deleted at the same time when systemd-tmpfiles decides to delete
them.